requirements.txt
torch>=1.10.0
transformers>=4.0.0
Pillow>=8.0.0
re
gc